+++ packages/core/src/currentScopes.ts
@@ -3,7 +3,7 @@ import { getGlobalSingleton, getMainCarrier } from './carrier';
 import type { Client } from './client';
 import { Scope } from './scope';
 import type { TraceContext } from './types-hoist';
-import { generateSpanId } from './utils-hoist';
+import { generateSpanId } from './utils-hoist/propagationContext';
 
 /**
  * Get the currently active scope.
