+++ packages/core/src/index.ts
@@ -121,8 +121,154 @@ export { _INTERNAL_captureLog, _INTERNAL_flushLogsBuffer } from './logs/exports'
 export { consoleLoggingIntegration } from './logs/console-integration';
 
 // TODO: Make this structure pretty again and don't do "export *"
-export * from './utils-hoist/index';
-// TODO: Make this structure pretty again and don't do "export *"
 export * from './types-hoist/index';
 
 export type { FeatureFlag } from './featureFlags';
+
+export { applyAggregateErrorsToEvent } from './utils-hoist/aggregate-errors';
+export { getBreadcrumbLogLevelFromHttpStatusCode } from './utils-hoist/breadcrumb-log-level';
+export { getComponentName, getLocationHref, htmlTreeAsString } from './utils-hoist/browser';
+export { dsnFromString, dsnToString, makeDsn } from './utils-hoist/dsn';
+// eslint-disable-next-line deprecation/deprecation
+export { SentryError } from './utils-hoist/error';
+export { GLOBAL_OBJ } from './utils-hoist/worldwide';
+export type { InternalGlobal } from './utils-hoist/worldwide';
+export { addConsoleInstrumentationHandler } from './utils-hoist/instrument/console';
+export { addFetchEndInstrumentationHandler, addFetchInstrumentationHandler } from './utils-hoist/instrument/fetch';
+export { addGlobalErrorInstrumentationHandler } from './utils-hoist/instrument/globalError';
+export { addGlobalUnhandledRejectionInstrumentationHandler } from './utils-hoist/instrument/globalUnhandledRejection';
+export {
+  addHandler,
+  maybeInstrument,
+  resetInstrumentationHandlers,
+  triggerHandlers,
+} from './utils-hoist/instrument/handlers';
+export {
+  isDOMError,
+  isDOMException,
+  isElement,
+  isError,
+  isErrorEvent,
+  isEvent,
+  isInstanceOf,
+  isParameterizedString,
+  isPlainObject,
+  isPrimitive,
+  isRegExp,
+  isString,
+  isSyntheticEvent,
+  isThenable,
+  isVueViewModel,
+} from './utils-hoist/is';
+export { isBrowser } from './utils-hoist/isBrowser';
+export { CONSOLE_LEVELS, consoleSandbox, logger, originalConsoleMethods } from './utils-hoist/logger';
+export type { Logger } from './utils-hoist/logger';
+export {
+  addContextToFrame,
+  addExceptionMechanism,
+  addExceptionTypeValue,
+  checkOrSetAlreadyCaught,
+  getEventDescription,
+  parseSemver,
+  uuid4,
+} from './utils-hoist/misc';
+export { isNodeEnv, loadModule } from './utils-hoist/node';
+export { normalize, normalizeToSize, normalizeUrlToBase } from './utils-hoist/normalize';
+export {
+  addNonEnumerableProperty,
+  convertToPlainObject,
+  // eslint-disable-next-line deprecation/deprecation
+  dropUndefinedKeys,
+  extractExceptionKeysForMessage,
+  fill,
+  getOriginalFunction,
+  markFunctionWrapped,
+  objectify,
+} from './utils-hoist/object';
+export { basename, dirname, isAbsolute, join, normalizePath, relative, resolve } from './utils-hoist/path';
+export { makePromiseBuffer, SENTRY_BUFFER_FULL_ERROR } from './utils-hoist/promisebuffer';
+export type { PromiseBuffer } from './utils-hoist/promisebuffer';
+export { severityLevelFromString } from './utils-hoist/severity';
+export {
+  UNKNOWN_FUNCTION,
+  createStackParser,
+  getFramesFromEvent,
+  getFunctionName,
+  stackParserFromStackParserOptions,
+  stripSentryFramesAndReverse,
+} from './utils-hoist/stacktrace';
+export { filenameIsInApp, node, nodeStackLineParser } from './utils-hoist/node-stack-trace';
+export { isMatchingPattern, safeJoin, snipLine, stringMatchesSomePattern, truncate } from './utils-hoist/string';
+export {
+  isNativeFunction,
+  supportsDOMError,
+  supportsDOMException,
+  supportsErrorEvent,
+  supportsFetch,
+  supportsHistory,
+  supportsNativeFetch,
+  supportsReferrerPolicy,
+  supportsReportingObserver,
+} from './utils-hoist/supports';
+export { SyncPromise, rejectedSyncPromise, resolvedSyncPromise } from './utils-hoist/syncpromise';
+export { browserPerformanceTimeOrigin, dateTimestampInSeconds, timestampInSeconds } from './utils-hoist/time';
+export {
+  TRACEPARENT_REGEXP,
+  extractTraceparentData,
+  generateSentryTraceHeader,
+  propagationContextFromHeaders,
+} from './utils-hoist/tracing';
+export { getSDKSource, isBrowserBundle } from './utils-hoist/env';
+export type { SdkSource } from './utils-hoist/env';
+export {
+  addItemToEnvelope,
+  createAttachmentEnvelopeItem,
+  createEnvelope,
+  createEventEnvelopeHeaders,
+  createSpanEnvelopeItem,
+  envelopeContainsItemType,
+  envelopeItemTypeToDataCategory,
+  forEachEnvelopeItem,
+  getSdkMetadataForEnvelopeHeader,
+  parseEnvelope,
+  serializeEnvelope,
+} from './utils-hoist/envelope';
+export { createClientReportEnvelope } from './utils-hoist/clientreport';
+export {
+  DEFAULT_RETRY_AFTER,
+  disabledUntil,
+  isRateLimited,
+  parseRetryAfterHeader,
+  updateRateLimits,
+} from './utils-hoist/ratelimit';
+export type { RateLimits } from './utils-hoist/ratelimit';
+export {
+  MAX_BAGGAGE_STRING_LENGTH,
+  SENTRY_BAGGAGE_KEY_PREFIX,
+  SENTRY_BAGGAGE_KEY_PREFIX_REGEX,
+  baggageHeaderToDynamicSamplingContext,
+  dynamicSamplingContextToSentryBaggageHeader,
+  parseBaggageHeader,
+  objectToBaggageHeader,
+} from './utils-hoist/baggage';
+export {
+  getSanitizedUrlString,
+  parseUrl,
+  stripUrlQueryAndFragment,
+  parseStringToURLObject,
+  isURLObjectRelative,
+  getSanitizedUrlStringFromUrlObject,
+} from './utils-hoist/url';
+export {
+  eventFromMessage,
+  eventFromUnknownInput,
+  exceptionFromError,
+  parseStackFrames,
+} from './utils-hoist/eventbuilder';
+export { callFrameToStackFrame, watchdogTimer } from './utils-hoist/anr';
+export { LRUMap } from './utils-hoist/lru';
+export { generateTraceId, generateSpanId } from './utils-hoist/propagationContext';
+export { vercelWaitUntil } from './utils-hoist/vercelWaitUntil';
+export { SDK_VERSION } from './utils-hoist/version';
+export { getDebugImagesForResources, getFilenameToDebugIdMap } from './utils-hoist/debug-ids';
+export { escapeStringForRegex } from './utils-hoist/vendor/escapeStringForRegex';
